+++ cvxpy/settings.py
@@ -55,9 +55,9 @@ GLPK_MI = "GLPK_MI"
 ECOS = "ECOS"
 ECOS_BB = "ECOS_BB"
 SCS = "SCS"
-GUROBI_LIN = "GUROBI_LIN"
+GUROBI = "GUROBI"
 SOLVERS = [ECOS, ECOS_BB, CVXOPT, GLPK,
-           GLPK_MI, SCS, GUROBI_LIN]
+           GLPK_MI, SCS, GUROBI]
 
 # Map of constraint types.
 EQ, LEQ, SOC, SOC_EW, SDP, EXP, BOOL, INT = range(8)