+++ cvxpy/tests/test_examples.py
@@ -313,10 +313,10 @@ class TestExamples(BaseTest):
             print("optimal value with GLPK_MI:", prob.value)
             self.assertAlmostEqual(prob.value, 6)
 
-        if GUROBI_LIN in installed_solvers():
+        if GUROBI in installed_solvers():
             # Solve with Gurobi.
-            prob.solve(solver=GUROBI_LIN)
-            print("optimal value with GUROBI_LIN:", prob.value)
+            prob.solve(solver=GUROBI)
+            print("optimal value with GUROBI:", prob.value)
             self.assertAlmostEqual(prob.value, 6)
 
         print(installed_solvers())